Add `optim.rs` module and `fold` function for constant folding
This is literally the only rust code I've ever written :P

It's a direct translation of a Haskell version that I cooked-up first,
and then fuzzed using LeanCheck.

I'm sure it's probably not very idiomatic, and I get the sense that the
boxing/unboxing done here is probably unnecessary. One way it might be
made a bit syntactically cleaner is by changing some of the constructor
names and then getting rid of the `<type_name>::` prefixes. Not sure if
that's generally frowned-upon in rust.