Add Haskell version
Look in `Optim/src/Main.hs` if you're curious. I copied over the ADT
definitions and implemented a really naive `fold` function. Then I
implemented LeanCheck `Listable` instances which allow us to fuzz the
implementation using an enumerative testing scheme. We just need some
sort of predicate as a proxy for correctness. I wrote a very basic
`isFoldedProgram` function which just checks for subtrees of the form:

     Op
    /  \
  Int  Int

Which should never appear in a correctly-optimized AST. However, there
could be bugs in the predicate, so this still potentially suffers from
the test oracle problem.

However I at least felt like the implementation of the predicate was
somewhat simpler than the constant folding, so it has that going for it!